refactor: split socks4.rs into submodules
Split monolithic socks4.rs (652 lines) into smaller modules:
- socks4/mod.rs (20 lines) - exports and module declaration
- socks4/protocol.rs (202 lines) - protocol types and constants
- socks4/handler.rs (256 lines) - handler implementation
- socks4/request.rs (216 lines) - request handling

Total: 694 lines across 4 files (better organization)
Validation: cargo check ✅ cargo test ✅
